+++ packages/orq-rc/RELEASES.md
@@ -3669,3 +3669,13 @@ Based on:
 - [typescript v3.13.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.12 - packages/orq-rc
+
+## 2025-09-16 10:53:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.615.2 (2.698.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.13 - packages/orq-rc